+++ CHANGELOG.md
@@ -14,6 +14,16 @@ All notable changes to `laravel-data` will be documented in this file.
 - Allow creating data objects using `from` without parameters
 - Add support for a Dto and Resource object
 
+## 3.11.0 - 2023-12-21
+
+- Add support for data structure caching #622
+
+## 3.10.1 - 2023-12-04
+
+- Make ValidationPath Stringable
+- Fix PHPStan
+- Improve performance when optional peoperty exists (#612)
+
 ## 3.10.0 - 2023-12-01
 
 A fresh release after a month of vacation, enjoy!